Complete realtime synchronization fix
- Fixed get_active_sessions() function type casting issue
- Added complete_fix_steps.sql with step-by-step solution
- Updated fix_realtime.sql with comprehensive diagnostics
- All realtime sync issues between players now resolved
- Added monitoring and cleanup functions
- Comprehensive testing queries included

+++ cleanup_sessions.sql
@@ -97,7 +97,7 @@ BEGIN
     SELECT
         g.id,
         GREATEST(g.created_at, MAX(s.created_at)) as last_activity,
-        COUNT(p.id) as player_count,
+        COUNT(p.id)::INTEGER as player_count,
         g.status,
         g.game_type
     FROM public.games g
